safari-600.7-branch/LayoutTests/http/tests/websocket/tests/hybi/url-parsing-expected.txt

CONSOLE MESSAGE: Wrong url scheme for WebSocket http://127.0.0.1:8000/websocket/tests/hybi/null
CONSOLE MESSAGE: Invalid url for WebSocket ws://javascript:a
CONSOLE MESSAGE: Wrong url scheme for WebSocket http://127.0.0.1:8000/applet
CONSOLE MESSAGE: Wrong url scheme for WebSocket javascript:a
CONSOLE MESSAGE: WebSocket port 25 blocked
CONSOLE MESSAGE: URL has fragment component ws://127.0.0.1/path#
CONSOLE MESSAGE: URL has fragment component ws://127.0.0.1/path#fragment
Test WebSocket URL parsing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket() threw exception TypeError: Not enough arguments.
PASS new WebSocket(null) threw exception Error: SyntaxError: DOM Exception 12.
PASS new WebSocket("ws://javascript:a") threw exception Error: SyntaxError: DOM Exception 12.
PASS new WebSocket("/applet") threw exception Error: SyntaxError: DOM Exception 12.
PASS new WebSocket("javascript:a") threw exception Error: SyntaxError: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:25/") threw exception Error: SecurityError: DOM Exception 18.
PASS (new WebSocket("ws://127.0.0.1:8880/a/../websocket/tests/hybi/simple")).URL is "ws://127.0.0.1:8880/websocket/tests/hybi/simple"
PASS (new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/simple?")).URL is "ws://127.0.0.1:8880/websocket/tests/hybi/simple?"
PASS (new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/simple?k=v")).URL is "ws://127.0.0.1:8880/websocket/tests/hybi/simple?k=v"
PASS new WebSocket("ws://127.0.0.1/path#") threw exception Error: SyntaxError: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1/path#fragment") threw exception Error: SyntaxError: DOM Exception 12.
PASS successfullyParsed is true

TEST COMPLETE
